Update RentalLease_WithOpenZeppelinSecurity
Smart contract that allows a landlord to dynamically set the terms of a lease. Smart contract updates to change pay periods monthly to ensure payments are made for the correct corresponding month and allows the landlord to be able to assess late fee's. Tenant is able to make multiple and partial payments within a month.