Create SmallAirlinesAddsNewFlights.sol
A small airlines has the ability to create and add new flights to their business. The smart contract will dynamically store each flight independently, allowing for the creation of multiple flights, as many as they want, without the need of uploading a new smart contract for each new flight. Customers can purchase dynamically priced tickets based on the amount of tickets remaining. Customers that successfully purchase a ticket are added to the flight record and are stored as being in that flight for accounting and accountability purposes.