Create RentalCarCompanyContract.sol
This contract facilitates the ability of a rental car company owner to input all of their vehicles in to the smart contract and available to be rented dynamically. The contract allows the renter to rent a vehicle of their choice - if it is available, and after paying the deposit fee - and set their beginning and end date for the rental. At the end of the rental, the contract will check to see if the vehicle is returned on time as specified at the time of rental and add late fee's as necessary - determined by the amount of days late that the rental vehicle is being returned. Once returned and balance paid in full, the renter is returned their deposit and the Vehicle is reset and available for rent again.